The article is focusing on the requirement of a fashion illustrator
Fashion Illustration is divided into three distinctive categories. The sensualist depicts the work with the traditional form of processes and the materials. The technocrats employ the digital means of creating the images. And the Gamines and Sophisticates are playful in their work which reflects the whimsical references of the traditional forms of the fashion illustrations. Along a good hand on Illustrations, the need to learn digital Illustration is high as it gives you the scope to be quick to upload and send across and make changes to the demand and requirement